Obtain MCMC dimensions from an mcmc.list

Quickly query the number of burn-in samples, post-burnin, thinning, number of chains, etc. from a mcmc.list object.

post_dim(post, types = NULL)

Arguments

post
A mcmc.list object.

types
The dimension types to return. Must contain some of "burn", "post_burn", "thin", "chains", "nodes". Defaults to NULL, in which case all of these are returned.

Value

A numeric vector with named elements, which may contain:

burn: The burn-in period + adapting phase (per chain).

post_burn: The post-burn-in period (per chain).

thin: The thinning interval post-burn-in.

chains: The number of chains.

saved: The number of saved samples across all chains.

params: The number of nodes with MCMC samples.

All of these will be returned if types = NULL, a subset can be returned by specifying (for example) types = c("burn", "thin").

Note

If the post object was thinned after MCMC completed using post_thin(), then the "burn" and "thin" dimensions will be improperly calculated. post_thin() performs post-MCMC thinning of mcmc.list objects, and is solely for developing long-running post-processing code, so this is okay.

Examples

# load example mcmc.list
data(cjs)

# get all relevant dimensions
post_dim(cjs)

#>      burn post_burn      thin    chains     saved    params 
#>     11000     50000       200         2       500        21 

# get only the number of chains
post_dim(cjs, "chains")

#> [1] 2

# get the thinning and burn-in intervals
post_dim(cjs, c("burn", "thin"))

#>  burn  thin 
#> 11000   200