Remove nodes from mcmc.list

Just like post_subset(), but keep all nodes except those that match.

post_remove(post, params, ask = TRUE, auto_escape = TRUE)

Arguments

post
A mcmc.list object.

params
A vector of regular expressions specifying the nodes to match for removal. Accepts multi-element vectors to match more than one node at a time. See match_params() and vignette("pattern-matching") for more details.

ask
Prompt user for a response prior to removing nodes?

auto_escape
Automatically escape "[" and "]" characters? See match_params() for details.

Value

A mcmc.list, identical in all ways to the original except that nodes matched by the params argument are removed. If the user responds "no" to the question when ask = TRUE, post is returned unaltered.

Examples

# load example mcmc.list
data(cjs)

# get names of all nodes
get_params(cjs)

#> [1] "B0"     "sig_B0" "B1"     "sig_B1" "b0"     "b1"     "SIG"    "p"     

# remove the SIG nodes
new_cjs = suppressMessages(post_remove(cjs, "SIG", ask = FALSE))

# get names of new output
get_params(new_cjs)

#> [1] "B0"     "sig_B0" "B1"     "sig_B1" "b0"     "b1"     "p"